Convert map aggregations to declared decomposition
map_agg decomposes into itself and map_union; multimap_agg gets a
hidden merge over its map-of-arrays intermediate.
