+++ core/trino-main/src/main/java/io/trino/operator/aggregation/MapAggregationFunction.java
@@ -20,7 +20,7 @@ import io.trino.spi.function.AggregationFunction;
 import io.trino.spi.function.AggregationState;
 import io.trino.spi.function.BlockIndex;
 import io.trino.spi.function.BlockPosition;
-import io.trino.spi.function.CombineFunction;
+import io.trino.spi.function.Decomposition;
 import io.trino.spi.function.Description;
 import io.trino.spi.function.InputFunction;
 import io.trino.spi.function.OutputFunction;
@@ -47,16 +47,8 @@ public final class MapAggregationFunction
         state.add(key, keyPosition, value, valuePosition);
     }
 
-    @CombineFunction
-    public static void combine(
-            @AggregationState({"K", "V"}) MapAggregationState state,
-            @AggregationState({"K", "V"}) MapAggregationState otherState)
-    {
-        state.merge(otherState);
-    }
-
     @SqlNullable
-    @OutputFunction("map(K, V)")
+    @OutputFunction(value = "map(K, V)", decomposition = @Decomposition(partial = "map_agg", output = "map_union"))
     public static void output(@AggregationState({"K", "V"}) MapAggregationState state, BlockBuilder out)
     {
         state.writeAll((MapBlockBuilder) out);
