+++ core/trino-main/src/main/java/io/trino/operator/aggregation/multimapagg/MultimapMergeAggregation.java
@@ -11,8 +11,9 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package io.trino.operator.aggregation;
+package io.trino.operator.aggregation.multimapagg;
 
+import io.trino.spi.block.ArrayBlock;
 import io.trino.spi.block.Block;
 import io.trino.spi.block.BlockBuilder;
 import io.trino.spi.block.MapBlockBuilder;
@@ -21,42 +22,44 @@ import io.trino.spi.block.ValueBlock;
 import io.trino.spi.function.AggregationFunction;
 import io.trino.spi.function.AggregationState;
 import io.trino.spi.function.Decomposition;
-import io.trino.spi.function.Description;
 import io.trino.spi.function.InputFunction;
 import io.trino.spi.function.OutputFunction;
 import io.trino.spi.function.SqlNullable;
 import io.trino.spi.function.SqlType;
 import io.trino.spi.function.TypeParameter;
-import io.trino.spi.type.Type;
 
-@AggregationFunction("map_union")
-@Description("Aggregate all the maps into a single map")
-public final class MapUnionAggregation
+// merges multimap_agg intermediates, which are maps from key to the collected values
+@AggregationFunction(value = "multimap_agg$merge", isOrderSensitive = true, hidden = true)
+public final class MultimapMergeAggregation
 {
-    private MapUnionAggregation() {}
+    private MultimapMergeAggregation() {}
 
     @InputFunction
     @TypeParameter("K")
     @TypeParameter("V")
     public static void input(
-            @TypeParameter("V") Type valueType,
-            @AggregationState({"K", "V"}) MapAggregationState state,
-            @SqlType("map(K,V)") SqlMap value)
+            @AggregationState({"K", "V"}) MultimapAggregationState state,
+            @SqlType("map(K, array(V))") SqlMap value)
     {
         int rawOffset = value.getRawOffset();
         Block rawKeyBlock = value.getRawKeyBlock();
         Block rawValueBlock = value.getRawValueBlock();
 
         ValueBlock rawKeyValues = rawKeyBlock.getUnderlyingValueBlock();
-        ValueBlock rawValueValues = rawValueBlock.getUnderlyingValueBlock();
+        ArrayBlock rawValueValues = (ArrayBlock) rawValueBlock.getUnderlyingValueBlock();
         for (int i = 0; i < value.getSize(); i++) {
-            state.add(rawKeyValues, rawKeyBlock.getUnderlyingValuePosition(rawOffset + i), rawValueValues, rawValueBlock.getUnderlyingValuePosition(rawOffset + i));
+            int keyPosition = rawKeyBlock.getUnderlyingValuePosition(rawOffset + i);
+            Block values = rawValueValues.getArray(rawValueBlock.getUnderlyingValuePosition(rawOffset + i));
+            ValueBlock valueElements = values.getUnderlyingValueBlock();
+            for (int j = 0; j < values.getPositionCount(); j++) {
+                state.add(rawKeyValues, keyPosition, valueElements, values.getUnderlyingValuePosition(j));
+            }
         }
     }
 
     @SqlNullable
-    @OutputFunction(value = "map(K, V)", decomposition = @Decomposition(partial = "map_union", output = "map_union"))
-    public static void output(@AggregationState({"K", "V"}) MapAggregationState state, BlockBuilder out)
+    @OutputFunction(value = "map(K, array(V))", decomposition = @Decomposition(partial = "multimap_agg$merge", output = "multimap_agg$merge"))
+    public static void output(@AggregationState({"K", "V"}) MultimapAggregationState state, BlockBuilder out)
     {
         state.writeAll((MapBlockBuilder) out);
     }
